+++ crates/gosub_pipeline/src/layering/layer.rs
@@ -13,9 +13,9 @@ impl LayerId {
     }
 }
 
-impl AddAssign<i32> for LayerId {
-    fn add_assign(&mut self, rhs: i32) {
-        self.0 += rhs as u64;
+impl AddAssign<u64> for LayerId {
+    fn add_assign(&mut self, rhs: u64) {
+        self.0 += rhs;
     }
 }
